+++ samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/ComponentsViewModel.kt
@@ -12,7 +12,7 @@ import org.jetbrains.jewel.samples.showcase.components.Borders
 import org.jetbrains.jewel.samples.showcase.components.Buttons
 import org.jetbrains.jewel.samples.showcase.components.Checkboxes
 import org.jetbrains.jewel.samples.showcase.components.ChipsAndTrees
-import org.jetbrains.jewel.samples.standalone.view.component.Dropdowns
+import org.jetbrains.jewel.samples.showcase.components.Dropdowns
 import org.jetbrains.jewel.samples.standalone.view.component.Icons
 import org.jetbrains.jewel.samples.standalone.view.component.Links
 import org.jetbrains.jewel.samples.standalone.view.component.ProgressBar
